FutureEvalBlock: Evaluate asynchronously in a future
This allows the user to dispatch threads without locking up Vim, which
has poor native support for asynchronous processing.

It may make sense to wrap _all_ evaluations in a future, then either
return with the result or push the future into *futures*, thereby
capping editor lock-up time to a user-definable number of ms.